fix(ci): python-coverage-shards — split part_002 under first-party line limit
test_pr_monitor_operator_hints_part_002.py grew to 1644 lines and tripped the
MAX_FIRST_PARTY_FILE_LINES (1500) maintainability guardrail. Move the cohesive
_preserved_commit_already_on_remote no-op cluster and the atomic preserved-head
marker test into a new part_004, and drop the now-unused ProtectedScopeDiffError
import from part_002.
